backend/requirements.txt
accelerate==0.26.1
aiohttp==3.9.3
aiosignal==1.3.1
aiosmtplib==2.0.2
amqp==5.2.0
annotated-types==0.6.0
anthropic==0.36.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
billiard==4.2.1
bleach==6.1.0
blinker==1.7.0
branca==0.7.0
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cmdstanpy==1.2.0
comm==0.1.3
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.26.1
distlib==0.3.8
distro==1.9.0
Django==5.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==2.4.2
email-validator==2.1.0.post1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.91.0
fastapi-jwt-auth==0.5.0
fastapi-mail==1.4.1
fastjsonschema==2.19.0
filelock==3.13.1
fiona==1.9.5
Flask==3.0.2
folium==0.15.1
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
geopandas==0.14.2
h11==0.14.0
holidays==0.41
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.6.1
joblib==1.3.2
JPype1==1.5.0
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
kombu==5.4.2
konlpy==0.6.0
lxml==5.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.8
networkx==3.2.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.26.2
openai==1.12.0
openapi==1.1.0
opencv-python==4.8.0.74
openpyxl==3.1.2
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==4.1.0
plotly==5.20.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
prophet==1.1.5
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.4
pydantic-settings==2.1.0
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==1.7.1
pymongo==4.6.1
PyMySQL==1.1.0
pyparsing==3.1.1
pyproj==3.6.1
PyQt3D-Qt5==5.15.11
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
pyrsistent==0.20.0
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.6.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.4.3
QtPy==2.4.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.15.2
safetensors==0.4.2
schedule==1.2.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
selenium==4.17.2
Send2Trash==1.8.2
shapely==2.0.2
shiboken6==6.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
stanio==0.3.0
starlette==0.24.0
statsmodels==0.14.1
sympy==1.12
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.37.2
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.20.0
vine==5.1.0
virtualenv==20.25.0
visdom==0.2.4
wcwidth==0.2.12
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wsproto==1.2.0
xlrd==2.0.1
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
